+++ src/main/java/org/ezcode/codetest/application/community/dto/request/DiscussionCreateRequest.java
@@ -1,7 +1,7 @@
 package org.ezcode.codetest.application.community.dto.request;
 
 import org.ezcode.codetest.domain.community.model.Discussion;
-import org.ezcode.codetest.domain.submission.model.entity.Language;
+import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.ezcode.codetest.domain.problem.model.entity.Problem;
 import org.ezcode.codetest.domain.user.model.entity.User;
