+++ src/main/java/org/ezcode/codetest/domain/community/repository/DiscussionRepository.java
@@ -3,7 +3,7 @@ package org.ezcode.codetest.domain.community.repository;
 import java.util.Optional;
 
 import org.ezcode.codetest.domain.community.model.Discussion;
-import org.ezcode.codetest.domain.submission.model.entity.Language;
+import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
