+++ src/main/java/org/ezcode/codetest/infrastructure/persistence/repository/submission/jpa/LanguageJpaRepository.java
@@ -1,6 +1,6 @@
 package org.ezcode.codetest.infrastructure.persistence.repository.submission.jpa;
 
-import org.ezcode.codetest.domain.submission.model.entity.Language;
+import org.ezcode.codetest.domain.language.model.entity.Language;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LanguageJpaRepository extends JpaRepository<Language, Long> {
